adding getUrl
The method works, it gets the url as expected and behaves like the WebKit
gtk browser does.

I have some minor concerns this point about how non portable the code is.
I make assumptions about the encoding the characters and the endianess of
Windows which are not at all likely to change but which would result in
code breakage if they did.

Namely I assume that windows will always be little endian and that the
cef_string_t contains wchar strings that are 2 bytes wide and can be
decoded as UTF16.