Window event preliminary work
So far I have it so that a browser can optionally be provided. However I
do not have the show event firing. If I tell CEF not to create it's own
window, no callbacks are fired from CEF except for the
"on_render_process_created".  Thus I do not know where to fire the "show"
event. Should I be firing this myself? There does not appear to be any
kind of show callback in the CEF API as far as I can tell.

Additionally I suspect that most of this code is unnecessary as CEF issue
520 tells me that the popupFeatures are not yet actually populated. They
are currently all zeroed. Thus we may want to remove this code in a later
commit but we will definitely need it later.